requirements.txt
fire==0.7.0
pillow==11.1.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.17.0
returns==0.25.0
termcolor==2.5.0
typing_extensions==4.13.0
